-added generic service interface -check if entity is existing in update and delete methods, throws EntityNotFoundException otherwise

+++ web-client/trunk/src/main/java/com/cityopt/test/UserGroupProjectRepositoryTest.java
@@ -101,7 +101,7 @@ public class UserGroupProjectRepositoryTest {
 		
 		userGroupProjectRepository.saveAndFlush(usergroupproject);
 		
-		assertNotEquals(0, usergroupproject.getUsergroupprojectid());
+		assertNotEquals((Integer) 0, usergroupproject.getUsergroupprojectid());
 	}
 	
 	
@@ -132,7 +132,7 @@ public class UserGroupProjectRepositoryTest {
 		//Cascade is enable!
 		userGroupProjectRepository.saveAndFlush(usergroupproject);
 		
-		assertNotEquals(0, usergroupproject.getUsergroupprojectid());
+		assertNotEquals((Integer) 0, usergroupproject.getUsergroupprojectid());
 	}
 	
 	@Test
